+++ lib/models/config.dart
@@ -7,8 +7,9 @@
 /// administrator mode, and whether to include the app or create an installer file.
 ///
 /// This file provides methods to create a [Config] instance from JSON or directly
-/// from the `pubspec.yaml` file, as well as a method to convert the configuration
-/// attributes into environment variables for further use.
+/// from the `pubspec.yaml` file and custom config file if provided, as well as
+/// a method to convert the configuration attributes into environment variables
+/// for further use.
 library;
 
 import 'dart:io';
@@ -34,8 +35,11 @@ class Config {
   /// The unique identifier (UUID) for the app being packaged.
   final String id;
 
-  /// The pubspec file sourced for this configuration.
+  /// The pubspec file sourced for this configuration, used as a fallback source of values.
   final File pubspecFile;
+
+  /// The config file sourced for this configuration, used as the main source of values.
+  final File configFile;
 
   /// The global pubspec name attribute, same name of the exe generated from flutter build.
   final String pubspecName;
@@ -76,7 +80,7 @@ class Config {
   /// Whether the installer requires administrator privileges.
   final AdminMode admin;
 
-  /// The build type (debug or release).
+  /// The build type (release, profile, or debug).
   final BuildType type;
 
   /// Whether to include the app in the installer.
@@ -103,6 +107,7 @@ class Config {
   /// Creates a [Config] instance with default values.
   const Config({
     required this.pubspecFile,
+    required this.configFile,
     required this.dlls,
     required this.files,
     required this.buildArgs,
@@ -122,7 +127,7 @@ class Config {
     required this.signTool,
     required this.arch,
     required this.vcRedist,
-    this.type = BuildType.debug,
+    this.type = BuildType.release,
     this.app = true,
     this.installer = true,
   });
@@ -133,58 +138,61 @@ class Config {
   /// The name of the executable file that will be created.
   String get exeName => "$name.exe";
 
-  /// Creates a [Config] instance from a JSON map, typically read from `pubspec.yaml`.
+  /// Creates a [Config] instance from a JSON map, typically read from `pubspec.yaml` and a config file (if provided).
   ///
   /// Validates the configuration and exits with an error if invalid values are found.
   factory Config.fromJson(
-    Map<String, dynamic> json, {
+    Map<String, dynamic> json,
+    Map<String, dynamic> configJson, {
     required CliConfig cliConfig,
     required File pubspecFile,
+    required File configFile,
   }) {
-    if (json['inno_bundle'] is! Map<String, dynamic>) {
-      CliLogger.exitError("inno_bundle section is missing from pubspec.yaml.");
+    final configName =
+        configFile == pubspecFile ? "pubspec.yaml" : "config file";
+    if (configJson['inno_bundle'] is! Map<String, dynamic>) {
+      CliLogger.exitError("inno_bundle section is missing from $configName.");
     }
-    final Map<String, dynamic> inno = json['inno_bundle'];
+    final Map<String, dynamic> inno = configJson['inno_bundle'];
 
     if (inno['id'] is! String) {
       CliLogger.exitError(
-          "inno_bundle.id attribute is missing from pubspec.yaml. "
+          "inno_bundle.id attribute is missing from $configName. "
           "Run `dart run inno_bundle:guid` to generate a new one, "
-          "then put it in your pubspec.yaml.");
+          "then put it in your $configName.");
     } else if (!Uuid.isValidUUID(fromString: inno['id'])) {
-      CliLogger.exitError("inno_bundle.id from pubspec.yaml is not valid. "
+      CliLogger.exitError("inno_bundle.id from $configName is not valid. "
           "Run `dart run inno_bundle:guid` to generate a new one, "
-          "then put it in your pubspec.yaml.");
+          "then put it in your $configName.");
     }
     final String id = inno['id'];
 
     if (json['name'] is! String) {
-      CliLogger.exitError("name attribute is missing from pubspec.yaml.");
+      CliLogger.exitError("name attribute is missing from $configName.");
     }
     final String pubspecName = json['name'];
 
     if (inno['name'] != null && !validFilenameRegex.hasMatch(inno['name'])) {
-      CliLogger.exitError("inno_bundle.name from pubspec.yaml is not valid. "
+      CliLogger.exitError("inno_bundle.name from $configName is not valid. "
           "`${inno['name']}` is not a valid file name.");
     }
     final String name = inno['name'] ?? pubspecName;
 
     if ((cliConfig.appVersion ?? inno['version'] ?? json['version'])
         is! String) {
-      CliLogger.exitError("version attribute is missing from pubspec.yaml.");
+      CliLogger.exitError("version attribute is missing from $configName.");
     }
     final String version =
         cliConfig.appVersion ?? inno['version'] ?? json['version'];
 
     if ((inno['description'] ?? json['description']) is! String) {
-      CliLogger.exitError(
-          "description attribute is missing from pubspec.yaml.");
+      CliLogger.exitError("description attribute is missing from $configName.");
     }
     final String description = inno['description'] ?? json['description'];
 
     if ((inno['publisher'] ?? json['maintainer']) is! String) {
       CliLogger.exitError("maintainer or inno_bundle.publisher attributes are "
-          "missing from pubspec.yaml.");
+          "missing from $configName.");
     }
     final String publisher = inno['publisher'] ?? json['maintainer'];
 
@@ -194,7 +202,7 @@ class Config {
 
     if (inno['installer_icon'] != null && inno['installer_icon'] is! String) {
       CliLogger.exitError("inno_bundle.installer_icon attribute is invalid "
-          "in pubspec.yaml.");
+          "in $configName.");
     }
     final installerIcon = inno['installer_icon'] != null
         ? p.join(
@@ -211,12 +219,12 @@ class Config {
 
     if (inno['languages'] != null && inno['languages'] is! List) {
       CliLogger.exitError("inno_bundle.languages attribute is invalid "
-          "in pubspec.yaml, only a list of strings is allowed.");
+          "in $configName, only a list of strings is allowed.");
     }
     final languages = (inno['languages'] as List?)
             ?.map((l) {
-              final languageError = Language.validateConfig(l);
-              if (languageError != null) CliLogger.exitError(languageError);
+              final error = Language.validateConfig(l, configName: configName);
+              if (error != null) CliLogger.exitError(error);
               final language = Language.getByNameOrNull(l);
               if (language == null) return null;
               return language;
@@ -229,13 +237,13 @@ class Config {
         inno['admin'] is! bool &&
         inno['admin'] != "auto") {
       CliLogger.exitError("inno_bundle.admin attribute is invalid value "
-          "in pubspec.yaml");
+          "in $configName");
     }
     final admin = AdminMode.fromOption(inno['admin'] ?? true);
 
     if (inno['license_file'] != null && inno['license_file'] is! String) {
       CliLogger.exitError("inno_bundle.license_file attribute is invalid "
-          "in pubspec.yaml.");
+          "in $configName.");
     }
 
     final licenseFilePath = p.join(
@@ -249,6 +257,7 @@ class Config {
 
     final signToolError = SignTool.validateConfig(
       inno["sign_tool"],
+      configName: configName,
       signToolName: cliConfig.signToolName,
       signToolCommand: cliConfig.signToolCommand,
       signToolParams: cliConfig.signToolParams,
@@ -261,7 +270,8 @@ class Config {
       signToolParams: cliConfig.signToolParams,
     );
 
-    final archError = BuildArch.validateConfig(inno['arch']);
+    final archError =
+        BuildArch.validateConfig(inno['arch'], configName: configName);
     if (archError != null) CliLogger.exitError(archError);
     final arch = BuildArch.fromOption(inno['arch']);
 
@@ -269,19 +279,19 @@ class Config {
         inno['vc_redist'] is! bool &&
         inno['vc_redist'] != "download") {
       CliLogger.exitError("inno_bundle.vc_redist attribute is invalid value "
-          "in pubspec.yaml");
+          "in $configName");
     }
     final vcRedist = VcRedistMode.fromOption(inno['vc_redist'] ?? true);
 
     if (inno['dlls'] != null && inno['dlls'] is! List) {
       CliLogger.exitError("inno_bundle.dlls attribute is invalid "
-          "in pubspec.yaml, only a list of dll entries is allowed.");
+          "in $configName, only a list of dll entries is allowed.");
     }
     final dlls = ((inno['dlls'] ?? []) as List)
         .map((d) {
           if (d == null) return null;
 
-          final dllError = DllEntry.validateConfig(d);
+          final dllError = DllEntry.validateConfig(d, configName: configName);
           if (dllError != null) CliLogger.exitError(dllError);
 
           return DllEntry.fromJson(d);
@@ -303,6 +313,7 @@ class Config {
 
     return Config(
       pubspecFile: pubspecFile,
+      configFile: configFile,
       buildArgs: cliConfig.buildArgs,
       id: id,
       pubspecName: pubspecName,
@@ -328,16 +339,24 @@ class Config {
     );
   }
 
-  /// Creates a [Config] instance directly from the `pubspec.yaml` file.
+  /// Creates a [Config] instance directly from the `pubspec.yaml` file and a config file (if provided).
   ///
   /// Provides a convenient way to load configuration without manual JSON parsing.
-  factory Config.fromFile(File pubspecFile, CliConfig cliConfig) {
-    final json = readPubspec(pubspecFile);
+  factory Config.fromFile(
+    File pubspecFile,
+    File configFile,
+    CliConfig cliConfig,
+  ) {
+    final pubspecJson = readYaml(pubspecFile);
+    final configJson =
+        configFile == pubspecFile ? pubspecJson : readYaml(configFile);
 
     return Config.fromJson(
-      json,
+      pubspecJson,
+      configJson,
       cliConfig: cliConfig,
       pubspecFile: pubspecFile,
+      configFile: configFile,
     );
   }
 
@@ -346,6 +365,7 @@ class Config {
     final variables = <String, String>{
       'APP_ID': id,
       'PUBSPEC_NAME': pubspecName,
+      'CONFIG_FILE': configFile.path,
       'APP_NAME': name,
       'APP_NAME_CAMEL_CASE': camelCase(name),
       'APP_DESCRIPTION': description,